extracted errors outside
