+++ core/util/merkle_tree.go
@@ -3,7 +3,7 @@ package util
 import (
 	"fmt"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 )
 
 /*MerkleTree - A data structure that implements MerkleTreeI interface */
