+++ core/zcncrypto/zcncrypto.go
@@ -7,7 +7,7 @@ import (
 	"github.com/0chain/gosdk/core/encryption"
 	"github.com/tyler-smith/go-bip39"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 )
 
 const CryptoVersion = "1.0"
