+++ go.mod
@@ -13,11 +13,11 @@ require (
 	go.dedis.ch/kyber/v3 v3.0.5
 	go.uber.org/zap v1.15.0
 	golang.org/x/crypto v0.0.0-20210322153248-0c34fe9e7dc2
-	github.com/0chain/gosdk/core/common/errors v0.0.0
+	github.com/0chain/gosdk/errors v0.0.0
 )
 
 go 1.13
 
 replace (
-	github.com/0chain/gosdk/core/common/errors => ./core/common/errors
+	github.com/0chain/gosdk/errors => ./errors
 )
