+++ zboxcore/fileref/list.go
@@ -1,7 +1,7 @@
 package fileref
 
 import (
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 	"github.com/mitchellh/mapstructure"
 )
