+++ zboxcore/marker/writemarker.go
@@ -3,8 +3,8 @@ package marker
 import (
 	"fmt"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
 	"github.com/0chain/gosdk/core/encryption"
+	zchainErrors "github.com/0chain/gosdk/errors"
 	"github.com/0chain/gosdk/zboxcore/client"
 )
