+++ zboxcore/sdk/filemetaworker_test.go
@@ -14,7 +14,7 @@ import (
 	"sync"
 	"testing"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 
 	"github.com/0chain/gosdk/core/zcncrypto"
 	"github.com/0chain/gosdk/zboxcore/blockchain"
