+++ zboxcore/sdk/renameworker.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"time"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 
 	"github.com/0chain/gosdk/zboxcore/fileref"
