+++ zboxcore/sdk/sdk.go
@@ -10,7 +10,7 @@ import (
 	"strconv"
 	"time"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 	"github.com/pkg/errors"
 
 	"github.com/0chain/gosdk/core/logger"
