+++ zboxcore/sdk/sharerequest.go
@@ -5,7 +5,7 @@ import (
 	b64 "encoding/base64"
 	"encoding/json"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 
 	"github.com/0chain/gosdk/core/common"
 	"github.com/0chain/gosdk/zboxcore/blockchain"
