+++ zboxcore/sdk/sync.go
@@ -13,7 +13,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 	"github.com/pkg/errors"
 
 	"github.com/0chain/gosdk/zboxcore/fileref"
