+++ zboxcore/sdk/uploadworker.go
@@ -16,7 +16,7 @@ import (
 	"os"
 	"sync"
 
-	zchainErrors "github.com/0chain/gosdk/core/common/errors"
+	zchainErrors "github.com/0chain/gosdk/errors"
 
 	"github.com/0chain/gosdk/core/util"
 	"github.com/0chain/gosdk/zboxcore/allocationchange"
